+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Deprecated
 ### Removed
 ### Fixed
+- Fixed issue #37  
+When multiple mouse buttons changed state in one update of the mouse, only one event would be fired for one of the buttons
 
 ## [0.4.1] - 2019-05-15
 ### Fixed